chore: reorganize database types and rename db to db_init
